consolidate cohere embed and openai embed
