+++ tests/entrypoints/pooling/embed/test_online_long_text.py
@@ -15,7 +15,7 @@ import pytest
 import pytest_asyncio
 
 from tests.utils import RemoteOpenAIServer
-from vllm.entrypoints.pooling.embed.protocol import EmbeddingResponse
+from vllm.entrypoints.pooling.embed.openai_protocol import EmbeddingResponse
 from vllm.platforms import current_platform
