+++ vllm/entrypoints/pooling/embed/openai_protocol.py
